chore: shared 디렉토리의 상수, 타입 관리 방식 변경
- feature 단위로 묶지않고 상수, 타입끼리 묶기
