+++ apps/frontend/src/shared/constants/weekdays.ts
